[NOJIRA] Refactor RN buttons
